+++ NEMODU/NEMODU/Screen/Map/ViewController/BottomSheet/FriendProfileBottomSheet.swift
@@ -374,14 +374,14 @@ extension FriendProfileBottomSheet {
     
     
     func bindToastMessage() {
-        viewModel.output.requestStatus
+        viewModel.requestStatus
             .withUnretained(self)
             .subscribe(onNext: { owner, status in
                 owner.popupToast(toastType: status ? .postFriendRequest : .networkError)
             })
             .disposed(by: disposeBag)
         
-        viewModel.output.deleteStatus
+        viewModel.deleteStatus
             .withUnretained(self)
             .subscribe(onNext: { owner, status in
                 owner.popupToast(toastType: status ? .cancelFriendRequest : .networkError)
